adding decimal functionionality

+++ server/routes/math.js
@@ -6,8 +6,8 @@ var path = require("path");
 router.get("/:operator/:x/:y", function(req,res){
     console.log("post received: ", req.params.operator, "y: ", req.params.y, "x: ", req.params.x )
     var answer = 0;
-    var y = parseInt(req.params.y);
-    var x = parseInt(req.params.x);
+    var y = parseFloat(req.params.y);
+    var x = parseFloat(req.params.x);
     switch (req.params.operator) {
       case "add":
         answer =  y + x
